+++ apps/autotest/tsql3.sh
@@ -1,50 +1,58 @@
 #!/bin/bash
 
-# Reverse proxy configuration
-if [ -z "$UB_RP_CONFIG" ]; then
-  export UB_RP_CONFIG=$PWD/rp-config-disable.json
-fi
-
 # Host configuration (set UB_HOST in case of reverse proxy)
 if [ -z "$UB_HOST" ]; then
-  export UB_HOST=http://localhost:8881
+  export UB_HOST='http://localhost:8881'
 fi
 
+if [ -z "$UB_APP" ]; then
+  export UB_APP='autotest'
+fi
+
+if [ -z "$UB_APPDATA" ]; then
+  export UB_APPDATA='./'
+fi
+
+export PLATFORM=lin
+
 err() {
-  echo Testcase $TESTCASE failed
+  echo Testcase "$TESTCASE" failed
   exit 1
 }
 
-rm -f ./_autotestResults*.json
-rm -f ./last_result.log
-rm ./logs/*.log
+rm -f "$UB_APPDATA"_autotestResults*.json
+rm -f "$UB_APPDATA"last_result.log
+rm "$UB_APPDATA"logs/*.log
 
 TESTCASE='hello'
 ub -e "console.log('Start autotest')"
 
-TESTCASE='drop database'
-rm -f ./*.sqlite3* || err
+TESTCASE='Check config'
+ub -T > /dev/null || err
 
-rm -rf ./documents/simple || err
+TESTCASE='drop database'
+rm -f "$UB_APPDATA"localdb/*.sqlite3* || err
+
+rm -rf "$UB_APPDATA"stores/documents/simple || err
 
 TESTCASE='init database'
 # Check whether UB_CFG set and contains a value, not spaces
-if [ -z ${UB_CFG+x} ] || [ -z "${UB_CFG// }" ]; then
+if [ -z "$UB_CFG" ]; then
   UB_CFG=ubConfig.json
 fi
 # Check whether UB_DBA set and contains a value, not spaces
-if [ -z ${UB_DBA+x} ] || [ -z "${UB_DBA// }" ]; then
+if [ -z "$UB_DBA" ]; then
   UB_DBA=sa
 fi
 # Check whether UB_DBAPWD set and contains a value, not spaces
-if [ -z ${UB_DBAPWD+x} ] || [ -z "${UB_DBAPWD// }" ]; then
+if [ -z "$UB_DBAPWD" ]; then
   UB_DBAPWD=sa
 fi
 # export UB_DEV=true
 
 npx ubcli createStore -cfg $UB_CFG -noLogo || err
 
-PASSWORD_FOR_ADMIN=admin
+PASSWORD_FOR_ADMIN='admin'
 npx ubcli initDB -cfg $UB_CFG -dba $UB_DBA -dbaPwd $UB_DBAPWD -p $PASSWORD_FOR_ADMIN -drop -create || err
 
 TESTCASE=generateDDL